Move collection cards into the Library
