+++ src/NexusMods.App.UI/NexusMods.App.UI.csproj
@@ -607,9 +607,6 @@
         <Compile Update="Pages\LibraryPage\Collections\CollectionCardDesignViewModel.cs">
           <DependentUpon>ICollectionCardViewModel.cs</DependentUpon>
         </Compile>
-        <Compile Update="Pages\LibraryPage\Collections\CollectionsDesignViewModel.cs">
-          <DependentUpon>ICollectionsViewModel.cs</DependentUpon>
-        </Compile>
         <Compile Update="Pages\CollectionDownload\CollectionDownloadDesignViewModel.cs">
           <DependentUpon>ICollectionDownloadViewModel.cs</DependentUpon>
         </Compile>
