+++ src/NexusMods.App.UI/Pages/CollectionDownload/CollectionDownloadViewModel.cs
@@ -123,8 +123,8 @@ public sealed class CollectionDownloadViewModel : APageViewModel<ICollectionDown
             {
                 var pageData = new PageData
                 {
-                    FactoryId = CollectionsPageFactory.StaticId,
-                    Context = new CollectionsPageContext
+                    FactoryId = LibraryPageFactory.StaticId,
+                    Context = new LibraryPageContext()
                     {
                         LoadoutId = targetLoadout,
                     },
