--- src/NexusMods.App.UI/Pages/LibraryPage/Collections/CollectionsView.axaml
@@ -1,37 +1,0 @@
-<reactiveUi:ReactiveUserControl x:TypeArguments="collections:ICollectionsViewModel"
-    xmlns="https://github.com/avaloniaui"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:reactiveUi="http://reactiveui.net"
-    xmlns:collections="clr-namespace:NexusMods.App.UI.Pages.LibraryPage.Collections"
-    xmlns:panels="clr-namespace:Avalonia.Labs.Panels;assembly=Avalonia.Labs.Panels"
-    mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-    x:Class="NexusMods.App.UI.Pages.LibraryPage.Collections.CollectionsView">
-    <Design.DataContext>
-        <collections:CollectionCardDesignViewModel />
-    </Design.DataContext>
-    
-    <ItemsControl x:Name="Collections">
-        <ItemsControl.ItemsPanel>
-            <ItemsPanelTemplate>
-                <panels:FlexPanel
-                    AlignContent="FlexStart"
-                    RowSpacing="{StaticResource Spacing-3}"
-                    ColumnSpacing="{StaticResource Spacing-3}"
-                    Wrap="Wrap"
-                    Margin="24" />
-            </ItemsPanelTemplate>
-        </ItemsControl.ItemsPanel>
-        <ItemsControl.ItemTemplate>
-            <DataTemplate>
-                <reactiveUi:ViewModelViewHost ViewModel="{CompiledBinding .}"
-                                              Width="508"
-                                              Height="288"
-                                              ClipToBounds="False"/>
-            </DataTemplate>
-        </ItemsControl.ItemTemplate>
-    </ItemsControl>
-
-</reactiveUi:ReactiveUserControl>
-
