--- src/NexusMods.App.UI/Pages/LibraryPage/Collections/CollectionsViewModel.cs
@@ -1,56 +1,0 @@
-using System.Collections.ObjectModel;
-using DynamicData;
-using NexusMods.Abstractions.NexusModsLibrary.Models;
-using NexusMods.App.UI.Windows;
-using NexusMods.App.UI.WorkspaceSystem;
-using NexusMods.Icons;
-using NexusMods.MnemonicDB.Abstractions;
-using ReactiveUI;
-using System.Reactive.Disposables;
-using NexusMods.Abstractions.Loadouts;
-
-namespace NexusMods.App.UI.Pages.LibraryPage.Collections;
-
-public class CollectionsViewModel : APageViewModel<ICollectionsViewModel>, ICollectionsViewModel
-{
-    public CollectionsViewModel(
-        IServiceProvider serviceProvider,
-        IConnection conn,
-        IWindowManager windowManager,
-        LoadoutId targetLoadout) : base(windowManager)
-    {
-        TabIcon = IconValues.ModLibrary;
-        TabTitle = "Collections (WIP)";
-
-        var loadout = Loadout.Load(conn.Db, targetLoadout);
-        var gameId = loadout.Installation.GameId;
-
-        this.WhenActivated(d =>
-        {
-            var tileImagePipeline = ImagePipelines.GetCollectionTileImagePipeline(serviceProvider);
-            var userAvatarPipeline = ImagePipelines.GetUserAvatarPipeline(serviceProvider);
-
-            CollectionMetadata.ObserveAll(conn)
-                .FilterImmutable(collection =>
-                {
-                    if (!CollectionMetadata.GameId.TryGetValue(collection, out var collectionGameId)) return true;
-                    return collectionGameId == gameId;
-                })
-                .Transform(ICollectionCardViewModel (coll) => new CollectionCardViewModel(
-                    tileImagePipeline: tileImagePipeline,
-                    userAvatarPipeline: userAvatarPipeline,
-                    windowManager: WindowManager,
-                    workspaceId: WorkspaceId,
-                    connection: conn,
-                    revision: coll.Revisions.First().RevisionId,
-                    targetLoadout: targetLoadout)
-                )
-                .Bind(out _collections)
-                .Subscribe()
-                .DisposeWith(d);
-        });
-    }
-
-    private ReadOnlyObservableCollection<ICollectionCardViewModel> _collections = new(new ObservableCollection<ICollectionCardViewModel>());
-    public ReadOnlyObservableCollection<ICollectionCardViewModel> Collections => _collections;
-}
