+++ src/NexusMods.App.UI/Pages/LibraryPage/ILibraryViewModel.cs
@@ -1,4 +1,6 @@
+using System.Collections.ObjectModel;
 using Avalonia.Platform.Storage;
+using NexusMods.App.UI.Pages.LibraryPage.Collections;
 using NexusMods.App.UI.WorkspaceSystem;
 using R3;
 
@@ -7,6 +9,7 @@ namespace NexusMods.App.UI.Pages.LibraryPage;
 public interface ILibraryViewModel : IPageViewModelInterface
 {
     LibraryTreeDataGridAdapter Adapter { get; }
+    ReadOnlyObservableCollection<ICollectionCardViewModel> Collections { get; }
 
     string EmptyLibrarySubtitleText { get; }
